changed author_rating to float

+++ server/migrations/20230707192707_create_table_threads.js
@@ -24,7 +24,7 @@ exports.up = function (knex) {
       .comment('0:Not Completed|1:Completed');
 
     table.text('body').nullable();
-    table.tinyint('author_rating').unsigned().defaultTo(0);
+    table.float('author_rating').unsigned().defaultTo(0);
     table.integer('total_upvotes').unsigned().defaultTo(0);
     table.integer('total_downvotes').unsigned().defaultTo(0);
     table.timestamps(true, true);
